Use dataclass-scalafix for code gen
This uses https://github.com/hamnis/dataclass-scalafix.
