+++ modules/definitions/src/main/scala/lmcoursier/credentials/Credentials.scala
@@ -1,1 +1,7 @@
-../../../../../../lm-coursier/src/main/scala/lmcoursier/credentials/Credentials.scala
+package lmcoursier.credentials
+
+import java.io.File
+
+abstract class Credentials extends Serializable
+
+object Credentials
