+++ modules/definitions/src/main/scala/lmcoursier/credentials/DirectCredentials.scala
@@ -6,17 +6,15 @@ import dataclass._
   host: String = "",
   username: String = "",
   password: String = "",
-  @since
+  @since("1.0")
   realm: Option[String] = None,
-  @since
+  @since("1.1")
   optional: Boolean = true,
-  @since
+  @since("1.2")
   matchHost: Boolean = false,
+  @since("1.3")
   httpsOnly: Boolean = true
 ) extends Credentials {
 
-  override def toString(): String =
-    withPassword("****")
-      .productIterator
-      .mkString("DirectCredentials(", ", ", ")")
+  override def toString(): String = s"DirectCredentials(host=$host, username=$username)"
 }
