+++ modules/lm-coursier/src/main/scala/lmcoursier/syntax/package.scala
@@ -182,23 +182,4 @@ package object syntax {
     def withRealm(realm: String): DirectCredentials =
       value.withRealm(Option(realm))
   }
-
-  implicit class CredentialsModule(value: Credentials.type) {
-    def apply(): DirectCredentials = DirectCredentials()
-    def apply(host: String, username: String, password: String): DirectCredentials =
-      DirectCredentials(host, username, password)
-    def apply(host: String, username: String, password: String, realm: Option[String]): DirectCredentials =
-      DirectCredentials(host, username, password, realm)
-    def apply(host: String, username: String, password: String, realm: String): DirectCredentials =
-      DirectCredentials(host, username, password, Option(realm))
-    def apply(host: String, username: String, password: String, realm: Option[String], optional: Boolean): DirectCredentials =
-      DirectCredentials(host, username, password, realm, optional, matchHost = false, httpsOnly = true)
-    def apply(host: String, username: String, password: String, realm: String, optional: Boolean): DirectCredentials =
-      DirectCredentials(host, username, password, Option(realm), optional, matchHost = false, httpsOnly = true)
-
-    def apply(f: File): FileCredentials =
-      FileCredentials(f.getAbsolutePath)
-    def apply(f: File, optional: Boolean): FileCredentials =
-      FileCredentials(f.getAbsolutePath, optional)
-  }
 }
